refactor modules
some fixes

more fixes + autoload models table

fix more bugs
