+++ src/download_base_models.py
@@ -1,14 +1,10 @@
-from pathlib import Path
 import requests
+from common import MDXNET_MODELS_DIR, RVC_MODELS_DIR
 
 MDX_DOWNLOAD_LINK = (
     "https://github.com/TRvlvr/model_repo/releases/download/all_public_uvr_models/"
 )
 RVC_DOWNLOAD_LINK = "https://huggingface.co/lj1995/VoiceConversionWebUI/resolve/main/"
-
-BASE_DIR = Path(__file__).resolve().parent.parent
-mdxnet_models_dir = BASE_DIR / "mdxnet_models"
-rvc_models_dir = BASE_DIR / "rvc_models"
 
 
 def dl_model(link, model_name, dir_name):
@@ -27,11 +23,11 @@ if __name__ == "__main__":
     ]
     for model in mdx_model_names:
         print(f"Downloading {model}...")
-        dl_model(MDX_DOWNLOAD_LINK, model, mdxnet_models_dir)
+        dl_model(MDX_DOWNLOAD_LINK, model, MDXNET_MODELS_DIR)
 
     rvc_model_names = ["hubert_base.pt", "rmvpe.pt"]
     for model in rvc_model_names:
         print(f"Downloading {model}...")
-        dl_model(RVC_DOWNLOAD_LINK, model, rvc_models_dir)
+        dl_model(RVC_DOWNLOAD_LINK, model, RVC_MODELS_DIR)
 
     print("All models downloaded!")
